perf: externalize typegpu, lazy MCPDiagnosticCard, consolidate formatBytes (-293KB)
D14: Externalize typegpu to CDN (esm.sh) — only used for a runtime
     WebGPU init probe in one panel, already behind dynamic import.
     Gate unplugin-typegpu to dev-only (apply: 'serve').

D17: Lazy-load MCPDiagnosticCard in ExecutionWorkspace and
     BatchProcessingPanel — only rendered on job failure.

D18: Consolidate 6 duplicate formatBytes implementations into a
     single shared export in src/lib/utils.ts.

D19: Remove dead @google/genai dependency (zero imports in source).

D20: Gate unplugin-typegpu plugin to serve-only (covered by D14).

Metric: client_js_bytes 1,648,669 → 1,355,301 (-293,368 bytes, -17.8%)
Cumulative from original: 2,655,341 → 1,355,301 (-48.9%)
All 1,414 tests pass (1,025 unit + 389 server). tsc clean. Lint clean.
