+++ index.html
@@ -10,7 +10,8 @@
       {
         "imports": {
           "onnxruntime-web": "https://esm.sh/onnxruntime-web@1.27.0",
-          "@huggingface/transformers": "https://esm.sh/@huggingface/transformers@4.2.0"
+          "@huggingface/transformers": "https://esm.sh/@huggingface/transformers@4.2.0",
+          "typegpu": "https://esm.sh/typegpu@0.11.9"
         }
       }
     </script>
