+++ package.json
@@ -52,7 +52,6 @@
     "tauri:build": "tauri build"
   },
   "dependencies": {
-    "@google/genai": "^2.15.0",
     "@huggingface/transformers": "^4.2.0",
     "@modelcontextprotocol/client": "^2.0.0",
     "@radix-ui/react-accordion": "^1.2.20",
