+++ src/components/features/assistant/LocalAiSetupCard.tsx
@@ -1,5 +1,6 @@
 import { Download, RefreshCw } from "lucide-react";
 import { openExternal } from "@/lib/openExternal";
+import { formatBytes } from "@/lib/utils";
 import { LocalModelManager } from "./LocalModelManager";
 import {
   LMS_STARTER_MODELS,
@@ -10,15 +11,6 @@ import {
   type LocalStarterModel,
 } from "./aiProviderCatalog";
 import type { LocalEngineSetup } from "./useLocalEngineSetup";
-
-/** Format bytes to human-readable size string. */
-function formatBytes(bytes: number): string {
-  if (bytes === 0) return "0 B";
-  const units = ["B", "KB", "MB", "GB", "TB"];
-  const i = Math.floor(Math.log(bytes) / Math.log(1024));
-  const size = bytes / Math.pow(1024, i);
-  return `${size.toFixed(i > 1 ? 1 : 0)} ${units[i]}`;
-}
 
 /** Best-effort match of a starter model tag against the engine's reported sizes. */
 function resolveDisplaySize(model: LocalStarterModel, modelSizes: Record<string, number>): string {
