+++ src/components/features/playground/WebGpuBenchmarkPanel.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useCallback, useRef, useEffect } from "react";
 import { Button, Label } from "@/components/ui";
-import { cn } from "@/lib/utils";
+import { cn, formatBytes } from "@/lib/utils";
 import {
   Globe,
   Cpu,
@@ -118,13 +118,6 @@ async function initWebGpuAdapter(): Promise<WebGpuStatus> {
   } catch (err) {
     return { available: false, adapterInfo: null, initError: String(err) };
   }
-}
-
-function formatBytes(bytes: number): string {
-  if (bytes >= 1024 * 1024 * 1024) return `${(bytes / (1024 * 1024 * 1024)).toFixed(1)} GB`;
-  if (bytes >= 1024 * 1024) return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-  if (bytes >= 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-  return `${bytes} B`;
 }
 
 function calcP50(arr: number[]): number {
