+++ vite.config.ts
@@ -1,4 +1,4 @@
-import typegpu from 'unplugin-typegpu/vite';
+import typegpuPlugin from 'unplugin-typegpu/vite';
 import react from '@vitejs/plugin-react';
 import tailwindcss from '@tailwindcss/vite';
 import path from 'path';
@@ -50,7 +50,9 @@ export default defineConfig(() => {
     plugins: [
       react(),
       tailwindcss(),
-      typegpu(),
+      // TypeGPU plugin: WGSL transpilation for dev DX only.
+      // In production, typegpu is externalized (loaded from CDN on-demand).
+      { ...typegpuPlugin(), apply: 'serve' },
       // Bundle analysis — open http://localhost:1420 to view tree map
       ...(process.env.ANALYZE
         ? [
@@ -115,11 +117,13 @@ export default defineConfig(() => {
         // Externalize heavy optional dependencies — they're only used in
         // Playground panels behind dynamic imports with graceful fallbacks.
         // The app works fully offline without them (Arena tokenizer falls back
-        // to prompt-derived encoding, Playground shows loading state).
+        // to prompt-derived encoding, Playground panels show loading state,
+        // WebGPU benchmark gracefully degrades without TypeGPU).
         // Users who open those panels load them from CDN on-demand.
         external: [
           '@huggingface/transformers',
           'onnxruntime-web',
+          'typegpu',
         ],
         output: {
           manualChunks(id: string) {
